update vestige header; fix 64 bit issues
git-svn-id: svn://localhost/ardour2/branches/3.0@10185 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ libs/ardour/ardour/session.h
@@ -744,10 +744,10 @@ class Session : public PBD::StatefulDestructible, public PBD::ScopedConnectionLi
 			
 	/*Native linuxVST support*/
 	
-	static long lxvst_callback (AEffect* effect,
-				  long opcode,
-				  long index,
-				  long value,
+	static intptr_t lxvst_callback (AEffect* effect,
+				  int32_t opcode,
+				  int32_t index,
+				  intptr_t value,
 				  void* ptr,
 				  float opt);
